% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\docType{data}
\name{testdata}
\alias{testdata}
\title{test dataset}
\format{Just the number \code{0}.}
\usage{
data(testdata)
}
\description{
a dataset for testing use of Roxygen2 to document datasets
}
\examples{
data(testdata)
}
\keyword{datasets}

